GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - dev (refs/gitbutler/dev)
   branch head: 9d2ca4a7de76e03755b2e145588bcd4e0ea43975
   - src/components/ui/dropdown-menu.tsx
   - src/app/room/[id]/page.tsx
   - src/components/ui/select.tsx
   - src/components/RoomMembers/RoomMembers.tsx
   - src/app/api/auth/[...nextauth]/route.ts
   - src/app/api/auth/[...nextauth]/option.ts
   - src/components/GuestAvatar/GuestAvatar.tsx
   - src/components/MemberAvatar/index.tsx
   - src/components/ui/switch.tsx
   - src/components/MemberAvatar/MemberAvatar.tsx
   - bun.lockb
   - src/components/NewRoom/NewRoom.tsx
   - src/middleware.ts
   - src/components/JoinRoomDialog/JoinRoomDialog.tsx
   - src/components/Room/types.ts
   - src/components/GuestAvatar/index.tsx
   - src/components/MemberAvatar/types.ts
   - README.md
   - src/components/Room/Room.tsx
   - src/components/JoinRoomDialog/types.ts
   - pnpm-lock.yaml
   - package.json
   - src/components/GuestAvatar/types.ts

Your previous branch was: refs/heads/dev

The sha for that commit was: 2164ad6ebc484bada064632130a75096b7c235ec

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch